+++ tactics2d/utils/__init__.py
@@ -1,0 +1,6 @@
+##! python3
+# Copyright (C) 2025, Tactics2D Authors. Released under the GNU GPLv3.
+# @File: __init__.py
+# @Description:
+# @Author: Tactics2D Team
+# @Version: